Add reset method for MappingFn to be used by tests.
Multiple tests run from the same jvm could be forced to reuse the same mappings leading to side-effects. To give tests the ability to re-initialize mappings, we expose this function that should only be called in a test class's @BeforeClass method.

Synchronizing on the static final object will necessitate that only one thread can be within this initialization code across all instances of this class. The others will wait for this thread to exit that block. Once the thread leaves the initialization flag will be set appropriately so it can't be initialized > 1 time unless the flag is reset.

The difference in functionality introduced by this change would be when multiple threads all want to initialize the mappings and an exception is thrown:
Before: All the other threads (aside from the one trying to initialize) will pass through normally without error and the mapping will remain uninitialized.
After: Each thread will try to initialize the mappings one by one, each throwing their own exception if unable.
PiperOrigin-RevId: 361817902
Change-Id: I337eb8607860a3cd1b026be389584190a61571fc